Move queries and mutations out of the Post containers
