+++ packages/client/src/modules/post/containers/Post.native.tsx
@@ -1,60 +1,10 @@
 import React from 'react';
-import { ApolloError } from 'apollo-client';
-import { graphql, compose, OptionProps } from 'react-apollo';
-import update from 'immutability-helper';
+import { compose } from 'react-apollo';
 
 import PostList from '../components/PostList.native';
-import { Post, PostProps, PostQueryResult, PostOperationResult } from '../types';
-import { Edge } from '../../../../../common/types';
+import { PostProps, PostQueryResult } from '../types';
 
-import POSTS_QUERY from '../graphql/PostsQuery.graphql';
-import POSTS_SUBSCRIPTION from '../graphql/PostsSubscription.graphql';
-import DELETE_POST from '../graphql/DeletePost.graphql';
-
-export function AddPost(prev: PostQueryResult, node: Post) {
-  // ignore if duplicate
-  if (prev.posts.edges.some((post: any) => node.id === post.cursor)) {
-    return prev;
-  }
-
-  const filteredPosts: Array<Edge<Post>> = prev.posts.edges.filter((post: any) => post.node.id !== null);
-
-  const edge: Edge<Post> = {
-    cursor: node.id,
-    node
-  };
-
-  return update(prev, {
-    posts: {
-      totalCount: {
-        $set: prev.posts.totalCount + 1
-      },
-      edges: {
-        $set: [edge, ...filteredPosts]
-      }
-    }
-  });
-}
-
-function DeletePost(prev: PostQueryResult, id: number) {
-  const index: number = prev.posts.edges.findIndex((x: Edge<Post>) => x.node.id === id);
-
-  // ignore if not found
-  if (index < 0) {
-    return prev;
-  }
-
-  return update(prev, {
-    posts: {
-      totalCount: {
-        $set: prev.posts.totalCount - 1
-      },
-      edges: {
-        $splice: [[index, 1]]
-      }
-    }
-  });
-}
+import { AddPost, DeletePost, withPostList, withPostDeleting, getSubscriptionOptions } from '../graphql';
 
 class PostComponent extends React.Component<PostProps, any> {
   public subscription: any;
@@ -91,20 +41,7 @@ class PostComponent extends React.Component<PostProps, any> {
   public subscribeToPostList = (endCursor: number) => {
     const { subscribeToMore } = this.props;
 
-    this.subscription = subscribeToMore({
-      document: POSTS_SUBSCRIPTION,
-      variables: { endCursor },
-      updateQuery: (prev: PostQueryResult, { subscriptionData: { data: { postsUpdated: { mutation, node } } } }) => {
-        let newResult: PostQueryResult = prev;
-        if (mutation === 'CREATED') {
-          newResult = AddPost(prev, node);
-        } else if (mutation === 'DELETED') {
-          newResult = DeletePost(prev, node.id);
-        }
-
-        return newResult;
-      }
-    });
+    this.subscription = subscribeToMore(getSubscriptionOptions(endCursor));
   };
 
   public render() {
@@ -112,63 +49,4 @@ class PostComponent extends React.Component<PostProps, any> {
   }
 }
 
-export default compose(
-  graphql<PostQueryResult>(POSTS_QUERY, {
-    options: () => {
-      return {
-        variables: { limit: 10, after: 0 }
-      };
-    },
-    props: ({ data }) => {
-      const { loading, error, posts, fetchMore, subscribeToMore } = data;
-      const loadMoreRows = () => {
-        return fetchMore({
-          variables: {
-            after: posts.pageInfo.endCursor
-          },
-          updateQuery: (previousResult, { fetchMoreResult }) => {
-            const totalCount = fetchMoreResult.posts.totalCount;
-            const newEdges = fetchMoreResult.posts.edges;
-            const pageInfo = fetchMoreResult.posts.pageInfo;
-
-            return {
-              // By returning `cursor` here, we update the `fetchMore` function
-              // to the new cursor.
-              posts: {
-                totalCount,
-                edges: [...previousResult.posts.edges, ...newEdges],
-                pageInfo,
-                __typename: 'Posts'
-              }
-            };
-          }
-        });
-      };
-      if (error) {
-        throw new ApolloError(error);
-      }
-      return { loading, posts, subscribeToMore, loadMoreRows };
-    }
-  }),
-  graphql(DELETE_POST, {
-    props: ({ mutate }: OptionProps<any, PostOperationResult>) => ({
-      deletePost: (id: number) => {
-        mutate({
-          variables: { id },
-          optimisticResponse: {
-            __typename: 'Mutation',
-            deletePost: {
-              id,
-              __typename: 'Post'
-            }
-          },
-          updateQueries: {
-            posts: (prev: PostQueryResult, { mutationResult: { data: { deletePost } } }) => {
-              return DeletePost(prev, deletePost.id);
-            }
-          }
-        });
-      }
-    })
-  })
-)(PostComponent);
+export default compose(withPostList, withPostDeleting)(PostComponent);
