+++ packages/client/src/modules/post/graphql/index.ts
@@ -1,0 +1,2 @@
+export * from './comments';
+export * from './posts';
