+++ packages/client/src/modules/post/types/comment.ts
@@ -1,7 +1,6 @@
 import { SubscribeToMoreOptions } from 'apollo-client';
+import { QueryProps } from 'react-apollo';
 import { Post } from './post';
-
-type SubscriptionToMoreFn = (option: SubscribeToMoreOptions) => void;
 
 // Comment types
 type AddCommentFn = (content: string, postId: number) => any;
@@ -9,11 +8,25 @@ type EditCommentFn = (id: number, content: string) => any;
 type DeleteCommentFn = (id: number) => any;
 type OnCommentSelectFn = (comment: Comment) => void;
 
+// Models
 interface Comment {
   id?: number;
   content: string;
 }
 
+interface PostCommentsProps extends QueryProps, CommentOperation {
+  postId: number;
+  comments: Comment[];
+  comment: Comment;
+}
+
+interface CommentProps {
+  postId: number;
+}
+
+export { Comment, PostCommentsProps, CommentProps}
+
+// Operations
 interface CommentOperation {
   addComment?: AddCommentFn;
   editComment?: EditCommentFn;
@@ -21,27 +34,15 @@ interface CommentOperation {
   onCommentSelect?: OnCommentSelectFn;
 }
 
-interface CommentQueryResult {
-  comment: Comment;
-}
-
-interface PostCommentsProps {
-  postId: number;
-  comments: Comment[];
-  comment: Comment;
-  onCommentSelect: OnCommentSelectFn;
-  subscribeToMore: SubscriptionToMoreFn;
-  addComment: AddCommentFn;
-  editComment: EditCommentFn;
-  deleteComment: DeleteCommentFn;
-}
-
-interface CommentProps {
-  postId: number;
-}
-
 interface CommentOperationResult {
   post: Post;
 }
 
-export { Comment, CommentOperation, CommentQueryResult, PostCommentsProps, CommentProps, CommentOperationResult };
+export { CommentOperation, CommentOperationResult }
+
+// Queries
+interface CommentQueryResult {
+  comment: Comment;
+}
+
+export { CommentQueryResult };
