+++ packages/client/src/modules/post/types/post.ts
@@ -1,7 +1,8 @@
 import { SubscribeToMoreOptions } from 'apollo-client';
 import { match as Match } from 'react-router';
+import { QueryProps } from 'react-apollo';
 import { Comment } from './comment';
-import { Location } from 'history';
+import { Location, History } from 'history';
 import { EntityList, PageInfo } from '../../../../../common/types';
 import { NavigationScreenProp } from 'react-navigation';
 
@@ -12,6 +13,9 @@ type EditPostFn = (id: number, title: string, content: string) => any;
 type DeletePostFn = (id: number) => any;
 type LoadMoreRowsFn = () => boolean;
 
+export { EditPostFn, AddPostFn, LoadMoreRowsFn };
+
+// Models
 interface Post {
   id: number;
   content: string;
@@ -24,21 +28,9 @@ interface Posts {
   pageInfo: PageInfo;
 }
 
-interface PostOperation {
-  addPost?: Post;
-  editPost?: (id: number, title: string, content: string) => void;
-}
-
-interface PostQuery {
-  post: Post;
-  loading: boolean;
-  subscribeToMore: (option: SubscribeToMoreOptions) => void;
-}
-
-interface PostEditProps {
+interface PostEditProps extends QueryProps {
   loading: boolean;
   post: Post;
-  subscribeToMore: (option: SubscribeToMoreOptions) => void;
   addPost: AddPostFn;
   editPost: EditPostFn;
   match: Match<any>;
@@ -47,45 +39,43 @@ interface PostEditProps {
   comments: Comment[];
 }
 
-interface PostProps {
+interface PostProps extends QueryProps {
   loadMoreRows: LoadMoreRowsFn;
   deletePost: DeletePostFn;
   loading: boolean;
   posts: EntityList<Post>;
-  subscribeToMore: (option: SubscribeToMoreOptions) => void;
   navigation: NavigationScreenProp<any>;
 }
 
-interface PostQueryResult {
-  loadMoreRows: () => EntityList<Post>;
-  loading: boolean;
-  subscribeToMore: (option: SubscribeToMoreOptions) => void;
-  posts: EntityList<Post>;
+export { Post, Posts, PostEditProps, PostProps };
+
+// Operations
+interface PostOperation {
+  addPost?: Post;
+  editPost?: (id: number, title: string, content: string) => void;
 }
 
 interface PostOperationResult {
-  deletePost: (id: number) => any;
-}
-
-interface PostListProps {
-  loading: boolean;
-  posts?: EntityList<Post>;
   deletePost: DeletePostFn;
-  loadMoreRows: LoadMoreRowsFn;
-  navigation: NavigationScreenProp<any>;
 }
 
-export {
-  Post,
-  Posts,
-  PostOperation,
-  PostQuery,
-  PostEditProps,
-  PostProps,
-  PostQueryResult,
-  PostOperationResult,
-  PostListProps,
-  LoadMoreRowsFn,
-  EditPostFn,
-  AddPostFn
-};
+interface PostEditOptionProps {
+  navigation?: NavigationScreenProp<any>;
+  history?: History;
+}
+
+export { PostOperation, PostOperationResult, PostEditOptionProps };
+
+// Queries
+interface PostQuery extends QueryProps {
+  post: Post;
+  loading: boolean;
+}
+
+interface PostQueryResult extends QueryProps {
+  loadMoreRows: () => EntityList<Post>;
+  loading: boolean;
+  posts: EntityList<Post>;
+}
+
+export { PostQuery, PostQueryResult };
